[MISC] Move counting_vector into separate header
To make review easier
